choose_disk: Detecting foreing disks
This patch is about considering if a disk that have ceph data is part of our cluster or not.
If the disk is not from our cluster, the disk cannot be said to be free
as we don't really know if the disk have important data on it, but we
cannot neither consider it to match our expectation.

This is the definition of a 'foreing' disk as in the Raid controllers.
We do recognize the metadata of it but cannot accept the disk in our
configuration.

This detection implies :
- testing first with ceph-volume
- testing then with ceph-disk

As ceph-disk will be deprecated, lets first get information from it to
avoid ceph-disk reporting ceph-volume's metadata. So this patch does
invert the detection order between both.

Then we try with ceph-disk. The detection is done once for all for the
system and then we parse it for every partition/block where we need to
consider if there is any metadata on it.

To consider if the disk is foreign, we compare the fsid from ansible
with the one detected by ceph-volume or ceph-disk.

This patch also add some protection around ceph-disk/ceph-volume & blkid
calls in case of the tool is not installed. Note that if blkid isn't
present, as it's the last way to detect any metadata, if it fails we
report a failure but continue. It would be dangerous to consider a disk
free while we don't really know its status.

This patch lack of detecting when ceph-disk reports an association between multiple devices.

A typical output looks like :

	############
	# Starting #
	############
	Detecting free devices with fsid=e8ba1fe2-dbf6-4b15-b0ca-2c5ffd97f59e
	Ignoring       dm-0 : device is a DM
	Ignoring        sda : device has existing partitions
	Adding         sdb1 : Ceph disk detected (journal)
	Adding         sdb2 : Ceph disk detected (journal)
	Adding         sdb3 : Ceph disk detected (journal)
	Adding         sdb4 : Ceph disk detected (journal)
	Adding         sdb5 : Ceph disk detected (block.db)
	Adding         sdb6 : Ceph disk detected (block.wal)
	Ignoring       sdc1 : device has foreign Ceph metadata :  a4a33288-6aa9-443d-a1aa-aa0fa655a057
	Adding         sdc2 : Ceph disk detected (block)
	Ignoring       sdd1 : device has foreign Ceph metadata :  a4a33288-6aa9-443d-a1aa-aa0fa655a057
	Adding         sdd2 : Ceph disk detected (block)
	Ignoring        sde : device has foreign Ceph metadata :  a4a33288-6aa9-443d-a1aa-aa0fa655a057
	Ignoring        sdf : read-only device
	Native syntax
	devices : {'storage_disks': {'count': 2, 'ceph_type': 'data', 'vendor': 'HP', 'size': 'gte(100 MB)'}}
	Looking for matches
	Inspecting storage_disks_000
	Inspecting storage_disks_001
	Matched devices   :   0
	Unmatched devices :   8
	  /dev/sdb1  (ceph journal  ) : [v:HP] [m:LOGICAL VOLUME] [s:5.00 GB] [r:0]
	  /dev/sdb2  (ceph journal  ) : [v:HP] [m:LOGICAL VOLUME] [s:5.00 GB] [r:0]
	  /dev/sdb3  (ceph journal  ) : [v:HP] [m:LOGICAL VOLUME] [s:5.00 GB] [r:0]
	  /dev/sdb4  (ceph journal  ) : [v:HP] [m:LOGICAL VOLUME] [s:5.00 GB] [r:0]
	  /dev/sdb5  (ceph block.db ) : [v:HP] [m:LOGICAL VOLUME] [s:1.00 GB] [r:0]
	  /dev/sdb6  (ceph block.wal) : [v:HP] [m:LOGICAL VOLUME] [s:576.00 MB] [r:0]
	  /dev/sdc2  (ceph block    ) : [v:HP] [m:LOGICAL VOLUME] [s:93.03 GB] [r:0]
	  /dev/sdd2  (ceph block    ) : [v:HP] [m:LOGICAL VOLUME] [s:279.27 GB] [r:1]
	### FATAL ###
	ERROR Could only find 0 of the 2 expected devices
	#############
	#######
	# End #
	#######

    # Veuillez saisir le message de validation pour vos modifications. Les lignes