choose_disk: check if the device is busy
This is the last thing we should be checking to consider a device can
used as a potential OSD.
There are applications such as databases that can consume a raw block
device, so before declaring the drive as usable for our OSD we must
ensure the drive is not accessed by any process.

For this, we try to open in readonly it using the os.O_EXCL flag. This
basically locks the device exclusively for the calling PID. If we can
lock it we get a fd, if not python will return an error saying the
device is busy.

Signed-off-by: Sébastien Han <seb@redhat.com>