choose_disk: Reject invalid disks
If a disk is improperly detected, the code is using a "failed" disk_label.
In that particular case, the disk should be ignored and not considered
as valid.

This patch change the logic of is_valid_partition_table() to become
is_invalid_partition_table() and return a "failed" if something is
wrong.

This patch also add implicit ignores of "failed" disks.

This patch also updates the tests accordingly.